Move generic tools in specific file #125
