+++ ansible_collections/arista/cvp/plugins/module_utils/tools_cv.py
@@ -92,65 +92,6 @@ def cv_connect(module):
     return client
 
 
-def isIterable(testing_object=None):
-    """
-    Test if an object is iterable or not.
-
-    Test if an object is iterable or not. If yes return True, else return False.
-
-    Parameters
-    ----------
-    testing_object : any, optional
-        Object to test if it is iterable or not, by default None
-
-    """
-    try:
-        some_object_iterator = iter(testing_object)  # noqa # pylint: disable=unused-variable
-        return True
-    except TypeError as te:  # noqa # pylint: disable=unused-variable
-        return False
-
-
-def match_filter(input, filter, default_always='all'):
-    """
-    Function to test if an object match userdefined filter.
-
-    Function support list of string and string as filter.
-    A default value is provided when calling function and if this default value for always matching is configured by user, then return True (Always matching)
-    If filter is a list, then we iterate over the input and check if it matches an entry in the filter.
-
-    Parameters
-    ----------
-    input : string
-        Input to test of that match filter or not.
-    filter : list
-        List of string to compare against input.
-    default_always : str, optional
-        Keyword to consider as always matching, by default 'all'
-    default_none : str, optional
-        Keyword to consider as never matching, by default 'none'
-
-    Returns
-    -------
-    bool
-        True if input matchs filter, False in other situation
-    """
-
-    # W102 Workaround to avoid list as default value.
-    if filter is None:
-        LOGGER.critical('Filter is not set, configure default value to [\'all\']')
-        filter = ["all"]
-
-    LOGGER.debug(" * is_in_filter - filter is %s", str(filter))
-    LOGGER.debug(" * is_in_filter - input string is %s", str(input))
-
-    if "all" in filter:
-        return True
-    elif any(element in input for element in filter):
-        return True
-    LOGGER.debug(" * is_in_filter - NOT matched")
-    return False
-
 
 def cv_update_configlets_on_device(module, device_facts, add_configlets, del_configlets):
     response = dict()
